refactor: extend WP_REST_Controller
