+++ src/API/REST/V3/Routes/Donations/DonationController.php
@@ -1,0 +1,368 @@
+<?php
+
+namespace Give\API\REST\V3\Routes\Donations;
+
+use Give\API\REST\V3\Routes\Donations\ValueObjects\DonationAnonymousMode;
+use Give\API\REST\V3\Routes\Donations\ValueObjects\DonationRoute;
+use Give\Donations\Models\Donation;
+use WP_Error;
+use WP_REST_Controller;
+use WP_REST_Request;
+use WP_REST_Response;
+use WP_REST_Server;
+
+/**
+ * @unreleased
+ */
+class DonationController extends WP_REST_Controller
+{
+    /**
+     * @var string
+     */
+    protected $namespace;
+
+    /**
+     * @var string
+     */
+    protected $rest_base;
+
+    /**
+     * @unreleased
+     */
+    public function __construct()
+    {
+        $this->namespace = DonationRoute::NAMESPACE;
+        $this->rest_base = DonationRoute::BASE;
+    }
+
+    /**
+     * @unreleased
+     */
+    public function register_routes()
+    {
+        register_rest_route($this->namespace, '/' . $this->rest_base . '/(?P<id>[\d]+)', [
+            [
+                'methods' => WP_REST_Server::READABLE,
+                'callback' => [$this, 'get_item'],
+                'permission_callback' => [$this, 'permissionsCheck'],
+                'args' => [
+                    'id' => [
+                        'type' => 'integer',
+                        'required' => true,
+                    ],
+                    'includeSensitiveData' => [
+                        'type' => 'boolean',
+                        'default' => false,
+                    ],
+                    'anonymousDonations' => [
+                        'type' => 'string',
+                        'default' => 'exclude',
+                        'enum' => [
+                            'exclude',
+                            'include',
+                            'redact',
+                        ],
+                    ],
+                ],
+            ],
+        ]);
+
+        register_rest_route($this->namespace, '/' . $this->rest_base, [
+            [
+                'methods' => WP_REST_Server::READABLE,
+                'callback' => [$this, 'get_items'],
+                'permission_callback' => [$this, 'permissionsCheck'],
+                'args' => $this->get_collection_params(),
+            ],
+        ]);
+    }
+
+    /**
+     * @unreleased
+     */
+    public function get_items($request)
+    {
+        $includeSensitiveData = $request->get_param('includeSensitiveData');
+        $donationAnonymousMode = new DonationAnonymousMode($request->get_param('anonymousDonations'));
+        $page = $request->get_param('page');
+        $perPage = $request->get_param('per_page');
+        $sortColumn = $this->getSortColumn($request->get_param('sort'));
+        $sortDirection = $request->get_param('direction');
+        $mode = $request->get_param('mode');
+        $status = $request->get_param('status');
+
+        $query = Donation::query();
+
+        if ($campaignId = $request->get_param('campaignId')) {
+            // Filter by CampaignId
+            $query->where('give_donationmeta_attach_meta_campaignId.meta_value', $campaignId);
+        }
+
+        if ($donorId = $request->get_param('donorId')) {
+            $query->where('give_donationmeta_attach_meta_donorId.meta_value', $donorId);
+        }
+
+        if ($donationAnonymousMode->isExcluded()) {
+            // Exclude anonymous donations from results
+            $query->where('give_donationmeta_attach_meta_anonymous.meta_value', 0);
+        }
+
+        // Include only current payment "mode"
+        $query->where('give_donationmeta_attach_meta_mode.meta_value', $mode);
+
+        // Filter by status if not 'any'
+        if ( ! in_array('any', (array)$status, true)) {
+            $query->whereIn('post_status', (array)$status);
+        }
+
+        $query
+            ->limit($perPage)
+            ->offset(($page - 1) * $perPage)
+            ->orderBy($sortColumn, $sortDirection);
+
+        $donations = $query->getAll() ?? [];
+        $donations = array_map(function ($donation) use ($includeSensitiveData, $donationAnonymousMode) {
+            return $this->escDonation($donation, $includeSensitiveData, $donationAnonymousMode);
+        }, $donations);
+
+        $totalDonations = empty($donations) ? 0 : Donation::query()->count();
+        $totalPages = (int)ceil($totalDonations / $perPage);
+
+        $response = rest_ensure_response($donations);
+        $response->header('X-WP-Total', $totalDonations);
+        $response->header('X-WP-TotalPages', $totalPages);
+
+        $base = add_query_arg(
+            map_deep($request->get_query_params(), function ($value) {
+                if (is_bool($value)) {
+                    $value = $value ? 'true' : 'false';
+                }
+
+                return urlencode($value);
+            }),
+            rest_url(DonationRoute::BASE)
+        );
+
+        if ($page > 1) {
+            $prevPage = $page - 1;
+
+            if ($prevPage > $totalPages) {
+                $prevPage = $totalPages;
+            }
+
+            $response->link_header('prev', add_query_arg('page', $prevPage, $base));
+        }
+
+        if ($totalPages > $page) {
+            $nextPage = $page + 1;
+            $response->link_header('next', add_query_arg('page', $nextPage, $base));
+        }
+
+        return $response;
+    }
+
+    /**
+     * @unreleased
+     */
+    public function get_item($request)
+    {
+        $donation = Donation::find($request->get_param('id'));
+        $includeSensitiveData = $request->get_param('includeSensitiveData');
+        $donationAnonymousMode = new DonationAnonymousMode($request->get_param('anonymousDonations'));
+
+        if ( ! $donation || ($donation->anonymous && $donationAnonymousMode->isExcluded())) {
+            return new WP_Error('donation_not_found', __('Donation not found', 'give'), ['status' => 404]);
+        }
+
+        return new WP_REST_Response($this->escDonation($donation, $includeSensitiveData, $donationAnonymousMode));
+    }
+
+    /**
+     * @unreleased
+     */
+    public function escDonation(
+        Donation $donation,
+        bool $includeSensitiveData = false,
+        DonationAnonymousMode $donationAnonymousMode = null
+    ): array {
+        $sensitiveDataExcluded = [];
+        if ( ! $includeSensitiveData) {
+            $sensitiveDataExcluded = [
+                'donorIp',
+                'email',
+                'phone',
+                'billingAddress',
+                'purchaseKey',
+            ];
+        }
+
+        $anonymousDataRedacted = [];
+        if ($donation->anonymous && $donationAnonymousMode->isRedacted()) {
+            $anonymousDataRedacted = [
+                'donorId',
+                'honorific',
+                'firstName',
+                'lastName',
+                'company',
+            ];
+        }
+
+        $donation = $donation->toArray();
+
+        foreach ($sensitiveDataExcluded as $property) {
+            if (array_key_exists($property, $donation)) {
+                unset($donation[$property]);
+            }
+        }
+
+        foreach ($anonymousDataRedacted as $property) {
+            if (array_key_exists($property, $donation)) {
+                $donation[$property] = __('anonymous', 'give');
+            }
+        }
+
+        return $donation;
+    }
+
+    /**
+     * @unreleased
+     */
+    public function getSortColumn(string $sortColumn): string
+    {
+        $sortColumnsMap = [
+            'id' => 'ID',
+            'createdAt' => 'post_date',
+            'updatedAt' => 'post_modified',
+            'status' => 'post_status',
+            'amount' => 'give_donationmeta_attach_meta_amount.meta_value',
+            'feeAmountRecovered' => 'give_donationmeta_attach_meta_feeAmountRecovered.meta_value',
+            'donorId' => 'give_donationmeta_attach_meta_donorId.meta_value',
+            'firstName' => 'give_donationmeta_attach_meta_firstName.meta_value',
+            'lastName' => 'give_donationmeta_attach_meta_lastName.meta_value',
+        ];
+
+        return $sortColumnsMap[$sortColumn];
+    }
+
+    /**
+     * @unreleased
+     */
+    public function get_collection_params(): array
+    {
+        $params = parent::get_collection_params();
+
+        $params['page']['default'] = 1;
+        $params['per_page']['default'] = 30;
+
+        // Remove default parameters not being used
+        unset($params['context']);
+        unset($params['search']);
+
+        $params += [
+            'sort' => [
+                'type' => 'string',
+                'default' => 'id',
+                'enum' => [
+                    'id',
+                    'createdAt',
+                    'updatedAt',
+                    'status',
+                    'amount',
+                    'feeAmountRecovered',
+                    'donorId',
+                    'firstName',
+                    'lastName',
+                ],
+            ],
+            'direction' => [
+                'type' => 'string',
+                'default' => 'DESC',
+                'enum' => ['ASC', 'DESC'],
+            ],
+            'mode' => [
+                'type' => 'string',
+                'default' => 'live',
+                'enum' => ['live', 'test'],
+            ],
+            'status' => [
+                'type' => 'array',
+                'items' => [
+                    'type' => 'string',
+                    'enum' => [
+                        'any',
+                        'publish',
+                        'give_subscription',
+                        'pending',
+                        'processing',
+                        'refunded',
+                        'revoked',
+                        'failed',
+                        'cancelled',
+                        'abandoned',
+                        'preapproval',
+                    ],
+                ],
+                'default' => ['any'],
+            ],
+            'campaignId' => [
+                'type' => 'integer',
+                'default' => 0,
+            ],
+            'donorId' => [
+                'type' => 'integer',
+                'default' => 0,
+            ],
+            'includeSensitiveData' => [
+                'type' => 'boolean',
+                'default' => false,
+            ],
+            'anonymousDonations' => [
+                'type' => 'string',
+                'default' => 'exclude',
+                'enum' => [
+                    'exclude',
+                    'include',
+                    'redact',
+                ],
+            ],
+        ];
+
+        return $params;
+    }
+
+    /**
+     * @unreleased
+     */
+    public function permissionsCheck(WP_REST_Request $request)
+    {
+        $isAdmin = current_user_can('manage_options');
+
+        $includeSensitiveData = $request->get_param('includeSensitiveData');
+        if ( ! $isAdmin && $includeSensitiveData) {
+            return new WP_Error(
+                'rest_forbidden',
+                esc_html__('You do not have permission to include sensitive data.', 'give'),
+                ['status' => $this->authorizationStatusCode()]
+            );
+        }
+
+        $donationAnonymousMode = new DonationAnonymousMode($request->get_param('anonymousDonations'));
+        if ( ! $isAdmin && $donationAnonymousMode->isIncluded()) {
+            return new WP_Error(
+                'rest_forbidden',
+                esc_html__('You do not have permission to include anonymous donations.', 'give'),
+                ['status' => $this->authorizationStatusCode()]
+            );
+        }
+
+        return true;
+    }
+
+    /**
+     * @unreleased
+     */
+    public function authorizationStatusCode(): int
+    {
+        return is_user_logged_in() ? 403 : 401;
+    }
+}
