+++ src/API/REST/V3/Routes/Donations/ValueObjects/DonationRoute.php
@@ -8,15 +8,12 @@ use Give\Framework\Support\ValueObjects\Enum;
  * @since 4.0.0
  *
  * @method static DonationRoute NAMESPACE()
- * @method static DonationRoute DONATION()
- * @method static DonationRoute DONATIONS()
+ * @method static DonationRoute BASE()
  * @method bool isNamespace()
- * @method bool isDonation()
- * @method bool isDonations()
+ * @method bool isBase()
  */
 class DonationRoute extends Enum
 {
     const NAMESPACE = 'givewp/v3';
-    const DONATION = 'donations/(?P<id>[0-9]+)';
-    const DONATIONS = 'donations';
+    const BASE = 'donations';
 }
